requirements/local.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
idna==2.9
lxml==4.5.0
mysqlclient==1.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
